Another refactoring of the build system and some bug fixes. Adds new storage containing the two new tables

+++ scripts/jail/cleanup.sh
@@ -2,10 +2,19 @@
 
 ROOT=$1
 
+SRCDIR=$2
+
 if [ "$ROOT" = "/" ]
 then
 	echo "Invalid argument"
 	exit 1
 fi
 
+if [ "$SRCDIR" = "/" ]
+then
+	echo "Invalid argument"
+	exit 1
+fi
+
 rm -rf $ROOT
+rm -rf $SRCDIR
